Move fetchDataFromJob to query-result
